refactor : 딜러에게 승패를 결정 짓는 책임 변경
