+++ src/main/kotlin/blackjack/domain/Participant.kt
@@ -1,7 +1,6 @@
 package blackjack.domain
 
 import blackjack.domain.enums.Condition
-import blackjack.domain.enums.MatchResult
 
 abstract class Participant(
     val name: String,
@@ -13,15 +12,5 @@ abstract class Participant(
 
     open fun hit(card: Card) {
         cards.append(card)
-    }
-
-    open fun determineResult(otherScore: Score): MatchResult {
-        val score = cards.calculateScore()
-
-        return when {
-            score.value > otherScore.value -> MatchResult.WIN
-            score.value < otherScore.value -> MatchResult.LOSE
-            else -> MatchResult.DRAW
-        }
     }
 }
